requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
idna==3.4
itsdangerous==2.1.2
loguru==0.7.0
magic-filter==1.0.9
MarkupSafe==2.1.3
multidict==6.0.4
openpyxl==3.1.2
peewee==3.16.2
pycparser==2.21
python-dotenv==1.0.0